requirements.txt
Django>=1.10,<1.11
Markdown==2.6.6
Pillow==3.3.0
django-filter>=1.0,<2
djangorestframework>=3.5,<4
